// src/orm/entities/transit/enums.ts
export enum RouteDirection {
  Forward = 'прямий',
  Reverse = 'зворотній',
}

export enum FineAppealStatus {
  Submitted = 'Подано',
  UnderReview = 'Перевіряється',
  Rejected = 'Відхилено',
  Approved = 'Прийнято',
}

export enum ComplaintStatus {
  Submitted = 'Подано',
  InReview = 'Розглядається',
  Resolved = 'Розглянуто',
}

export enum FineStatus {
  InProgress = 'В процесі',
  Paid = 'Оплачено',
  Cancelled = 'Відмінено',
  Overdue = 'Просрочено',
}